+++ release.sh
@@ -12,6 +12,13 @@ SVN="http://svn.km3net.de/auxiliaries/KM3Pipe"
 
 git checkout develop
 git pull
+git push
+
+git checkout master
+git pull
+
+git checkout develop
+git pull
 
 git flow release start "${VERSION}"